feat: Implement QuickSambaServer container
In this commit, most of the container's code has been implemented
based on existing code from the QuickFtpServer project. This allows
reading the configuration file and extracting the list of resources
and users configured therein. Additionally, a version of the
./docker-cmd.sh script has been included to facilitate testing and
debugging tasks (this script also originates from the other project).