feat: Enhance user experience for USER_ID/GROUP_ID setup
This commit refactors the way QuickSambaServer obtains the USER_ID
and GROUP_ID used for virtual users. The logic remains the same as
in the previous commit, but now provides more informative feedback
to the user.

If the user doesn't explicitly define USER_ID and GROUP_ID, the
script will now inform them that the values will be extracted
from the file permissions of the configuration file.